requirements.txt
# ML / DL Frameworks
torch==2.4.0
# torchaudio==2.4.0          
torchvision==0.19        

# Hugging Face & Timm
# accelerate
# diffusers
# timm
transformers

# ML/DL Helpers
einops==0.8.0

# Data Science
# matplotlib==3.9.3
opencv-python==4.10.0.84
pandas==2.2.3
scikit-image==0.22.0
scikit-learn==1.5.2
# sympy==1.13.3

# MLOps & Infrastructure
wandb==0.19.0 

# Utilities
# imageio==2.34.0 
PyYAML==6.0.2
tqdm==4.67.1
albumentations
ftfy
open-clip-torch